pipeline/requirements.txt
pandas
psycopg2-binary
python-dotenv
requests
pytest